@@ -8,13 +8,18 @@ class ReleasesController < ApplicationController
   # POST /release
   def create
     @last_release = @release = Release.current
-    @release = Release.new params[:release]
+    @release = Release.new release_params
     @release.save!
     @release.import_contexts_from @last_release
     @last_release.update_attribute :current, false
     redirect_to root_url
   rescue ActiveRecord::RecordInvalid
     render action: 'new'
+  end
+
+  protected
+  def release_params
+    params.require(:release).permit(:name)
   end
   
 end
