@@ -5,7 +5,7 @@ require 'rails/all'
 # If you have a Gemfile, require the default gems, the ones in the
 # current environment and also include :assets gems if in development
 # or test environments.
-Bundler.require *Rails.groups(:assets) if defined?(Bundler)
+Bundler.require(:default, Rails.env)
 
 module Chekov
   class Application < Rails::Application
@@ -36,8 +36,5 @@ module Chekov
 
     # Configure sensitive parameters which will be filtered from the log file.
     config.filter_parameters += [:password]
-
-    # Enable the asset pipeline
-    config.assets.enabled = true
   end
 end
